specs now ignore case when checking gem name
There was a failing test because the Arabic-Prawn gem is capitalized,
but the directory was lowercase. We should ignore case for directory
names (and keep them all lower case).

The spec now uses a case insensitive check to make sure that the `gem`
field in the yaml matches the directory.