Add language that links Trix to ActionText
Not all users of ActionText are aware that Trix is included in ActionText.
This change makes that relationship more obvious, so that users are not
confused as to why warnings for Trix are showing up for users of ActionText.

This also adds a link to the Rails forum post that describes this issue.